Creacion de niveles y refactor de codigo
